[Fix] metadata string building warning
The function to print all metadata-containing components
	metadata_table_print_all_components
contained a line which produced a compilation warning. This warning
might be harmless, but for correctness and to remove the warning the
line has been changed from
	... sprintf(line, "%s%s", line, known[i]);
to
	... strcat(line, known[i]);

The old version overwrote the `line` buffer by its own value and then
added the new value at its end. The new version achieves the same effect
without an overlapping buffer warning.